+++ e2e/shadow-dom/shadow-dom.spec.ts
@@ -11,4 +11,12 @@ test.describe('shadow-dom', () => {
     await page.focus('#input1');
     await expect(page.locator('#complete1')).toBeVisible();
   });
+
+  test('complete manually connected to input and anchor in shadow dom should work', async ({
+    page,
+  }) => {
+    await page.focus('#input2');
+    await expect(page.locator('#complete2')).toBeVisible();
+    expect((await page.locator('#complete2').boundingBox()).width).toBe(342);
+  });
 });